requirements.txt
click
jinja2>=3.0.3
pyyaml>=6.0
pycodestyle>=2.8.0
pandas>=1.4.0
isodate>=0.5.4
xarray>=0.11.3
questionary>=1.10.0
flake8>=6.0.0
